seems to be perfect

+++ util/file_decompressor.cpp
@@ -9,12 +9,13 @@ file_decompressor::file_decompressor(std::string file_name) : reader(file_name) 
         symbols_in_file = reader.get_n_bytes_r(8);
 
         auto converted_tree_size = static_cast<uint32_t>(reader.get_n_bytes(4));
-        if (converted_tree_size > SYMBOL_CNT*4) {
+        if (converted_tree_size > SYMBOL_CNT * 4) {
             throw;
         }
         auto converted_tree = std::vector<std::pair<int32_t, int32_t>>();
         for (uint32_t i = 0; i < converted_tree_size; ++i) {
-            converted_tree.emplace_back(static_cast<int32_t>(reader.get_n_bytes(4)), static_cast<int32_t>(reader.get_n_bytes(4)));
+            converted_tree.emplace_back(static_cast<int32_t>(reader.get_n_bytes(4)),
+                                        static_cast<int32_t>(reader.get_n_bytes(4)));
         }
         decompressor.set_tree(new code_tree(converted_tree));
     } catch (...) {
@@ -25,7 +26,7 @@ file_decompressor::file_decompressor(std::string file_name) : reader(file_name) 
 void file_decompressor::decompress(std::string dst) {
     writer.set_file(dst);
     //while (!reader.eof() || reader.has_useful_bits()) {
-    for (int i = 0; i < symbols_in_file; ++i) {
+    for (uint64_t i = 0; i < symbols_in_file; ++i) {
         writer.print(decompressor.decode(reader.get_next_code_piece()));
         reader.make_n_bits_used(decompressor.code_pos()); // code_pos() seems to be ok
     }
